+++ src/App.js
@@ -124,7 +124,7 @@ import getTickUniqueID from 'app-util/getTickUniqueID';
 import {navBarConfig, whiteNavBarConfig, routerConfig} from './navBarConfig';
 import {addJob} from './helper/jobsOnReset';
 import ItemAttribute from './components/stores/ItemAttribute';
-import InternetBankingModal from './components/payment/InternetBankingModal';
+import PaymentMethodDetailModal from './components/payment/PaymentMethodDetailModal';
 import AwesomeAlert from 'react-native-awesome-alerts';
 import Schedule, {ProductSchedule} from './containers/Schedule';
 import {ScheduleConfirm} from './containers/Schedule/Confirm';
@@ -161,6 +161,7 @@ import ModalConfirm from './components/ModalConfirm';
 import ProductStamps from './containers/ProductStamps';
 import ModalComboLocation from './components/ModalComboLocation';
 import APIHandler from './network/APIHandler';
+import Transaction from './components/payment/Transaction';
 
 /**
  * Not allow font scaling
@@ -1747,6 +1748,16 @@ class RootRouter extends Component {
               />
             </Lightbox>
 
+            {/* ================ MODAL WEBVIEW ================ */}
+            <Stack key={appConfig.routes.modalWebview}>
+              <Scene
+                key={`${appConfig.routes.modalWebview}_1`}
+                component={WebView}
+                renderBackButton={CloseButton}
+                back
+              />
+            </Stack>
+
             {/* ================ MODAL SHOW QR/BAR CODE ================ */}
             <Stack key={appConfig.routes.qrBarCode}>
               <Scene
@@ -1772,8 +1783,8 @@ class RootRouter extends Component {
             <Stack key={appConfig.routes.internetBanking}>
               <Scene
                 key={`${appConfig.routes.internetBanking}_1`}
-                title="Thẻ ATM"
-                component={InternetBankingModal}
+                // title="Thẻ ATM"
+                component={PaymentMethodDetailModal}
                 renderBackButton={() => <CloseButton color="#fff" />}
                 {...navBarConfig}
                 back
@@ -1795,6 +1806,16 @@ class RootRouter extends Component {
                 key={`${appConfig.routes.modalCameraView}_1`}
                 component={CameraView}
                 hideNavBar
+                />
+            </Stack>
+            
+            {/* ================ MODAL TRANSACTION================ */}
+            <Stack key={appConfig.routes.transaction} panHandlers={null}>
+              <Scene
+                key={`${appConfig.routes.transaction}_1`}
+                component={Transaction}
+                hideNavBar
+                back
               />
             </Stack>
           </Modal>