+++ src/components/RightButtonNavBar/RightButtonNavBar.tsx
@@ -43,9 +43,14 @@ class RightButtonNavBar extends Component<RightButtonNavBarProps> {
   state = {
     noti: 0,
   };
+  autoUpdateDisposer = () => {};
 
   componentDidMount() {
     this.updateNoti();
+  }
+
+  componentWillUnmount() {
+    this.autoUpdateDisposer();
   }
 
   get icon() {
@@ -154,7 +159,7 @@ class RightButtonNavBar extends Component<RightButtonNavBarProps> {
   }
 
   updateNoti() {
-    autorun(() => {
+    this.autoUpdateDisposer = autorun(() => {
       switch (this.props.type) {
         case RIGHT_BUTTON_TYPE.SHOPPING_CART:
           if (