+++ src/components/payment/Transaction/QRCodeFrame/index.js
@@ -1,0 +1,1 @@
+export {QRCodeFrame} from './QRCodeFrame';