+++ src/helper/handleStatusBarStyle.js
@@ -28,6 +28,8 @@ export default function handleStatusBarStyle(prevState, newState, action) {
 
     appConfig.routes.phoneAuth,
     appConfig.routes.qrBarCode,
+    appConfig.routes.modalWebview,
+    appConfig.routes.transaction,
 
     appConfig.routes.myVoucher,
     appConfig.routes.voucherDetail,